move database to the top level
