+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/data/FavoritesRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package com.example.android.playlistmaker.medialibrary.data
 
 import com.example.android.playlistmaker.medialibrary.data.converters.TrackConverter
-import com.example.android.playlistmaker.medialibrary.data.db.AppDatabase
-import com.example.android.playlistmaker.medialibrary.data.db.entity.TrackEntity
 import com.example.android.playlistmaker.medialibrary.domain.db.FavoritesRepository
 import com.example.android.playlistmaker.medialibrary.domain.models.Track
+import com.example.android.playlistmaker.util.data.db.AppDatabase
+import com.example.android.playlistmaker.util.data.db.entity.TrackEntity
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
