--- app/src/main/java/com/example/android/playlistmaker/medialibrary/data/db/dao/FavoriteTrackDao.kt
@@ -1,12 +1,0 @@
-package com.example.android.playlistmaker.medialibrary.data.db.dao
-
-import androidx.room.Dao
-import androidx.room.Query
-import com.example.android.playlistmaker.medialibrary.data.db.entity.TrackEntity
-
-@Dao
-interface FavoriteTrackDao {
-
-    @Query("SELECT * FROM favorite_tracks ORDER BY createdAt DESC;")
-    suspend fun getFavoriteTracks(): List<TrackEntity>
-}
