+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/di/DataModule.kt
@@ -1,7 +1,7 @@
 package com.example.android.playlistmaker.medialibrary.di
 
 import androidx.room.Room
-import com.example.android.playlistmaker.medialibrary.data.db.AppDatabase
+import com.example.android.playlistmaker.util.data.db.AppDatabase
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
