+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/domain/models/Track.kt
@@ -17,7 +17,7 @@ data class Track(
     val country: String,
     val previewUrl: String?,
     var isFavorite: Boolean = false,
-    val createdAt: Long
+    val createdAt: Long,
 ) : Parcelable {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
