+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/ui/PlaylistFragment.kt
@@ -1,8 +1,6 @@
 package com.example.android.playlistmaker.medialibrary.ui
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import com.example.android.playlistmaker.databinding.FragmentPlaylistBinding
