+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/ui/TrackAdapter.kt
@@ -1,7 +1,6 @@
 package com.example.android.playlistmaker.medialibrary.ui
 
 import android.view.ViewGroup
-import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.example.android.playlistmaker.medialibrary.domain.models.Track
