--- app/src/main/java/com/example/android/playlistmaker/player/data/db/entity/TrackEntity.kt
@@ -1,20 +1,0 @@
-package com.example.android.playlistmaker.player.data.db.entity
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "favorite_tracks")
-data class TrackEntity(
-    @PrimaryKey val trackId: Long,
-    val trackName: String,
-    val artistName: String,
-    val trackTime: Long,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val primaryGenreName: String,
-    val releaseDate: String,
-    val country: String,
-    val previewUrl: String?,
-    val createdAt: Long
-)
