+++ app/src/main/java/com/example/android/playlistmaker/player/di/DataModule.kt
@@ -3,8 +3,8 @@ package com.example.android.playlistmaker.player.di
 import android.media.MediaPlayer
 import androidx.room.Room
 import com.example.android.playlistmaker.player.data.PlayerClient
-import com.example.android.playlistmaker.player.data.db.AppDatabase
 import com.example.android.playlistmaker.player.data.player.AndroidStandardPlayerClient
+import com.example.android.playlistmaker.util.data.db.AppDatabase
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
