+++ app/src/main/java/com/example/android/playlistmaker/player/di/ViewModelModule.kt
@@ -8,7 +8,6 @@ import org.koin.dsl.module
 val playerViewModelModule = module {
 
     viewModel { (track: Track) ->
-//        PlayerViewModel(track, get(), get(), get())
         PlayerViewModel(track, get())
     }
 }
