+++ app/src/main/java/com/example/android/playlistmaker/util/data/db/AppDatabase.kt
@@ -1,9 +1,9 @@
-package com.example.android.playlistmaker.medialibrary.data.db
+package com.example.android.playlistmaker.util.data.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.android.playlistmaker.medialibrary.data.db.dao.FavoriteTrackDao
-import com.example.android.playlistmaker.medialibrary.data.db.entity.TrackEntity
+import com.example.android.playlistmaker.util.data.db.dao.FavoriteTrackDao
+import com.example.android.playlistmaker.util.data.db.entity.TrackEntity
 
 @Database(version = 1, entities = [TrackEntity::class], exportSchema = false)
 abstract class AppDatabase : RoomDatabase() {
