+++ app/src/main/java/com/example/android/playlistmaker/util/data/db/entity/TrackEntity.kt
@@ -1,6 +1,5 @@
-package com.example.android.playlistmaker.medialibrary.data.db.entity
+package com.example.android.playlistmaker.util.data.db.entity
 
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
@@ -16,5 +15,5 @@ data class TrackEntity(
     val releaseDate: String,
     val country: String,
     val previewUrl: String?,
-    val createdAt: Long
+    val createdAt: Long,
 )
