+++ app/src/main/java/com/example/android/playlistmaker/util/ui/BindingFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.viewbinding.ViewBinding
 
-abstract class BindingFragment<T: ViewBinding> : Fragment() {
+abstract class BindingFragment<T : ViewBinding> : Fragment() {
 
     private var _binding: T? = null
     protected val binding get() = _binding!!
